+++ lib/transforms/addRelNoopenerToBlankTargetAnchors.js
@@ -1,0 +1,31 @@
+module.exports = function (queryObj) {
+    return function addRelNoopenerToBlankTargetAnchors(assetGraph) {
+        var externalAnchors = assetGraph.findRelations({
+            type: 'HtmlAnchor',
+            hrefType: 'absolute'
+            // FIXME: also check that the relation is cross origin
+        }, true).filter(function (relation) {
+            return relation.node.getAttribute('target') === '_blank';
+        });
+
+        externalAnchors.forEach(function (relation) {
+            var node = relation.node;
+            var currentRel = node.getAttribute('rel');
+
+            if (typeof currentRel === 'string') {
+                var rels = currentRel.split(' ').map(function (str) {
+                    return str.toLowerCase();
+                });
+
+                if (rels.indexOf('noopener') === -1) {
+                    node.setAttribute('rel', rels.concat('noopener').join(' '));
+                }
+
+            } else {
+                relation.node.setAttribute('rel', 'noopener');
+            }
+
+            relation.from.markDirty();
+        });
+    };
+};
