+++ testdata/transforms/addRelNoopenerToBlankTargetAnchors/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Document</title>
+</head>
+<body>
+    <a href="index.html#">document fragment</a>
+    <a href="index.html#" target="_blank">document fragment in new window</a>
+    <a href="index.html#" target="_blank" rel="nofollow">document fragment in new window with rel=nofollow</a>
+    <a href="index.html#" target="_blank" rel="noopener">document fragment in new window with rel=noopener</a>
+
+    <a href="http://fisk.dk">cross domain link</a>
+    <a href="http://fisk.dk" target="_blank">cross domain link in new window</a>
+    <a href="http://fisk.dk" target="_blank" rel="nofollow">cross domain link in new window with rel=nofollow</a>
+    <a href="http://fisk.dk" target="_blank" rel="noopener">cross domain link in new window with rel=noopener</a>
+</body>
+</html>
